adding test cases
