+++ random_profile/assets/cities_name.txt
@@ -1,0 +1,27 @@
+Metropolis
+Eerie
+King's Landing
+Sunnydale
+Bedrock
+South Park
+Atlantis
+Mordor
+Olympus
+Dawnstar
+Balmora
+Gotham
+Springfield
+Quahog
+Smalltown
+Epicburg
+Pythonville
+Faketown
+Westworld
+Thundera
+Vice City
+Blackwater
+Oldtown
+Valyria
+Winterfell
+Braavos
+Lakeview
