+++ random_profile/assets/states_names.txt
@@ -1,0 +1,12 @@
+Main
+High
+Pearl
+Maple
+Park
+Oak
+Pine
+Cedar
+Elm
+Washington
+Lake
+Hill
