+++ tests/test_package.py
@@ -1,0 +1,35 @@
+import unittest
+import sys
+sys.path.append('.')
+from random_profile import RandomProfile
+
+random_profile = RandomProfile(num=1)
+
+class RandomProfileTest(unittest.TestCase):
+    def test_fname(self):
+        self.assertEqual(len(random_profile.first_name()), 1)
+    
+    def test_faname_with_num(self):
+        self.assertEqual(len(RandomProfile(num=10).first_name()), 10)
+    
+    def test_lname(self):
+        self.assertEqual(len(random_profile.last_name()), 1)
+    
+    def test_lname_with_num(self):
+        self.assertEqual(len(RandomProfile(num=10).last_name()), 10)
+    
+    def test_full_name(self):
+        self.assertEqual(len(random_profile.full_name()), 1)
+    
+    def test_full_profile_with_num(self):
+        self.assertEqual(len(RandomProfile(num=10).full_profile()), 10)
+    
+    def test_full_profile(self):
+        self.assertEqual(len(random_profile.full_profile()), 1)
+    
+    def test_full_profile_with_num(self):
+        self.assertEqual(len(RandomProfile(num=10).full_profile()), 10)
+    
+
+if __name__ == "__main__":
+    unittest.main()
